src/Pulsar.Runtime/Engine/SendMessageActionExecutor.cs
using System.Threading.Tasks;
using Pulsar.RuleDefinition.Models;
using Serilog;

namespace Pulsar.Runtime.Engine;

/// <summary>
/// Executes actions that send messages to external systems
/// </summary>
public class SendMessageActionExecutor : IActionExecutor
{
    private readonly ILogger _logger;

    public SendMessageActionExecutor(ILogger logger)
    {
        _logger = logger.ForContext<SendMessageActionExecutor>();
    }

    public Task<bool> ExecuteAsync(RuleAction action)
    {
        if (action.SendMessage == null || action.SendMessage.Count == 0)
        {
            return Task.FromResult(true);
        }

        try
        {
            foreach (var (channel, message) in action.SendMessage)
            {
                // TODO: Implement actual message sending logic based on channel
                // This could be email, SMS, webhook, etc.
                _logger.Information("Sending message to {Channel}: {Message}", channel, message);
            }

            return Task.FromResult(true);
        }
        catch (Exception ex)
        {
            _logger.Error(ex, "Failed to send messages");
            return Task.FromResult(false);
        }
    }
}
